src/requirements.txt
Flask==3.1.*
Flask-SQLAlchemy==3.1.*
Flask-Login==0.6.*
Flask-Bcrypt==1.0.*
Flask-WTF==1.2.*
PyMySQL==1.1.*
python-dotenv==1.0.*
